Create a socket simulator that is instantiated on a URL param
