+++ assets/js/app.js
@@ -1,18 +1,25 @@
 var $ = require('jquery');
 var _ = require('underscore');
+var socketio = require('socket.io-client');
 
 var view = require('./views/centre');
 
 $(function() {
   var options = {};
   var params = document.location.search.replace("?","").split("&");
-  var endpointParam = _.find(params, function(item) {
-    return item.match('endpoint');
+
+  var simulatorParam = _.find(params, function(item) {
+    return item.match('simulator');
   });
 
-  if (endpointParam) {
-    options.url = endpointParam.split("=")[1];
+  if (simulatorParam) {
+    var simulator = require('./socketSimulator');
+    socketio = simulator.client;
+
+    simulator.start();
   }
 
-  new view();
+  options.socket = socketio;
+
+  new view(options);
 });
